refactor: split markdown preview actions
